renderer_opengl: isolate core presentation code
